Moved multiscale object to TileJSON
