+++ docs/index.md
@@ -67,25 +67,3 @@ A FeatureCollection object is a JSON object representing a collection of feature
     - `"type"`: A string with the value “StitchingVector”
 
     Any object of a StitchingVector “features” array MUST be an “Image” special type of features object.
-
-### Multiscale Object
-
-A multiscale object represents the choice of axes (2-5D) and potentially their transformations that should be applied to the numerical data in order to arrive to the actual size of the object described. It MUST have the following properties:
-
-- `"axes"`: Representing the choice of axes as an array of Axis objects.
-
-It may contain either of, but NOT both of the following properties:
-
-- `"coordinateTransformations"`: Representing the set of coordinate transformations that should be applied to the numerical data in order to arrive to the actual size of the object described. It MUST be an array of objects, each object representing a coordinate transformation. Each object MUST have properties as follows:
-  - `"type"`: Representing the type of the coordinate transformation. Currently supported types are `"identity"`, `"scale"`, and `"translate"`. If the type is `"scale"`, the object MUST have the property `"scale"`, representing the scaling factor. It MUST be an array of numbers, with the number of elements equal to the number of axes in the coordinate system. If the type is `"translate"`, the object MUST have the property `"translate"`, representing the translation vector. It MUST be an array of numbers, with the number of elements equal to the number of axes in the coordinate system. If the type is `"identity"`, the object MUST NOT have any other properties.
-- `"transformationMatrix"`: Representing the transformation matrix from the coordinate system of the image to the coordinate system of the MicroJSON object. It MUST be an array of arrays of numbers, with the number of rows equal to the number of axes in the coordinate system, and the number of columns equal to the number of axes in the image coordinate system. The transformation matrix MUST be invertible.
-
-### Axis Object
-
-Together with the other axes in the axes array, an axis object represents the coordinate system of the MicroJSON object (2D-5D)
-It MUST have the following properties:
-
-- `"name"`: Representing the name of the axis. It MUST be a string.
-It may contain the following properties:
-- `"unit"`: Representing the units of the corresponding axis of the geometries in the MicroJSON object. It MUST be an array with the elements having any of the following values: `[“angstrom", "attometer", "centimeter", "decimeter", "exameter", "femtometer", "foot", "gigameter", "hectometer", "inch", "kilometer", "megameter", "meter", "micrometer", "mile", "millimeter", "nanometer", "parsec", "petameter", "picometer", "terameter", "yard", "yoctometer", "yottameter", "zeptometer", "zettameter“]`
-- `"description"`: A string describing the axis.
