+++ src/microjson/tilemodel.py
@@ -1,5 +1,7 @@
-from typing import List, Optional, Union, Dict
+from typing import List, Optional, Union, Dict, Literal
+from enum import Enum
 from pydantic import BaseModel, AnyUrl, conlist, RootModel
+from pydantic import StrictStr
 from pathlib import Path
 
 
@@ -27,6 +29,119 @@ class TileLayer(BaseModel):
     fieldranges: Optional[Dict[str, List[Union[int, float, str]]]] = None
     fieldenums: Optional[Dict[str, List[str]]] = None
     fielddescriptions: Optional[Dict[str, str]] = None
+
+
+class Unit(Enum):
+    """A unit of measurement"""
+
+    ANGSTROM = "angstrom"
+    ATTOMETER = "attometer"
+    CENTIMETER = "centimeter"
+    DECIMETER = "decimeter"
+    EXAMETER = "exameter"
+    FEMTOMETER = "femtometer"
+    FOOT = "foot"
+    GIGAMETER = "gigameter"
+    HECTOMETER = "hectometer"
+    INCH = "inch"
+    KILOMETER = "kilometer"
+    MEGAMETER = "megameter"
+    METER = "meter"
+    MICROMETER = "micrometer"
+    MILE = "mile"
+    MILLIMETER = "millimeter"
+    NANOMETER = "nanometer"
+    PARSEC = "parsec"
+    PETAMETER = "petameter"
+    PICOMETER = "picometer"
+    TERAMETER = "terameter"
+    YARD = "yard"
+    YOCTOMETER = "yoctometer"
+    YOTTAMETER = "yottameter"
+    ZEPTOMETER = "zeptometer"
+    ZETTAMETER = "zettameter"
+    PIXEL = "pixel"
+    RADIAN = "radian"
+    DEGREE = "degree"
+
+
+class AxisType(Enum):
+    """The type of an axis"""
+
+    SPACE = "space"
+    TIME = "time"
+    CHANNEL = "channel"
+
+
+class Axis(BaseModel):
+    """An axis of a coordinate system
+
+    Args:
+        name (StrictStr): The name of the axis
+        type (Optional[AxisType]): The type of the axis
+        unit (Optional[Unit]): The unit of the axis
+        description (Optional[str]): A description of the axis
+    """
+
+    name: StrictStr
+    type: Optional[AxisType] = None
+    unit: Optional[Unit] = None
+    description: Optional[str] = None
+
+
+class CoordinateTransformation(BaseModel):
+    """Coordinate transformation abstract class
+    harmonized with the OME model"""
+
+
+class Identity(CoordinateTransformation):
+    """Identity transformation for coordinates
+
+    Args:
+        type (Literal["identity"]): The type of the transformation
+    """
+
+    type: Literal["identity"] = "identity"
+
+
+class Translation(CoordinateTransformation):
+    """Translation transformation
+
+    Args:
+        type (Literal["translation"]): The type of the transformation
+        translation (List[float]): The translation vector
+    """
+
+    type: Literal["translation"] = "translation"
+    translation: List[float]
+
+
+class Scale(CoordinateTransformation):
+    """Scale transformation
+
+    Args:
+        type (Literal["scale"]): The type of the transformation
+        scale (List[float]): The scale vector
+    """
+
+    type: Literal["scale"] = "scale"
+    scale: List[float]
+
+
+class Multiscale(BaseModel):
+    """A coordinate system for MicroJSON coordinates
+
+    Args:
+        axes (List[Axis]): The axes of the coordinate system
+        coordinateTransformations (Optional[List[CoordinateTransformation]]):
+            A list of coordinate transformations
+        transformationMatrix (Optional[List[List[float]]):
+            The transformation matrix
+    """
+
+    axes: List[Axis]
+    coordinateTransformations: Optional[List[CoordinateTransformation]] = None
+    transformationMatrix: Optional[List[List[float]]] = None
 
 
 class TileModel(BaseModel):
@@ -78,6 +193,8 @@ class TileModel(BaseModel):
         max_length=6)] = None
     fillzoom: Optional[int] = None
     vector_layers: List[TileLayer]
+    multiscale: Optional[Multiscale] = None
+    scale_factor: Optional[float] = None
 
 
 class TileJSON(RootModel):
