--- tests/json/microjson/invalid/empty-coordinatesystem.json
@@ -1,25 +1,0 @@
-{ 
-  "type": "Feature",
-  "geometry": {
-    "type": "Polygon",
-    "coordinates": [
-      [
-        [100.0, 0.0],
-        [101.0, 0.0],
-        [101.0, 1.0],
-        [100.0, 1.0],
-        [100.0, 0.0]
-      ],
-      [
-        [100.8, 0.8],
-        [100.8, 0.2],
-        [100.2, 0.2],
-        [100.2, 0.8],
-        [100.8, 0.8]
-      ]
-    ]
-  }
-  ,
-  "multiscale": {
-  }
-}
