+++ tests/json/microjson/valid/3d.json
@@ -15,27 +15,5 @@
       [100.2, 0.8, 1],
       [100.8, 0.8, 1]
     ]
-  ],
-  "multiscale": {
-    "axes": [
-      {
-          "name": "x",
-          "unit": "micrometer",
-          "type": "space",
-          "description": "x-axis"
-      },
-      {
-          "name": "y",
-          "unit": "micrometer",
-          "type": "space",
-          "description": "y-axis"
-      },
-      {
-          "name": "z",
-          "unit": "micrometer",
-          "type": "space",
-          "description": "z-axis"
-      }
   ]
-  }
 }
