--- tests/json/microjson/valid/axesonly.json
@@ -1,35 +1,0 @@
-{
-  "type": "Polygon",
-  "coordinates": [
-    [
-      [100.0, 0.0],
-      [101.0, 0.0],
-      [101.0, 1.0],
-      [100.0, 1.0],
-      [100.0, 0.0]
-    ],
-    [
-      [100.8, 0.8],
-      [100.8, 0.2],
-      [100.2, 0.2],
-      [100.2, 0.8],
-      [100.8, 0.8]
-    ]
-  ],
-  "multiscale": {
-    "axes": [
-      {
-          "name": "x",
-          "unit": "micrometer",
-          "type": "space",
-          "description": "x-axis"
-      },
-      {
-          "name": "y",
-          "unit": "micrometer",
-          "type": "space",
-          "description": "y-axis"
-      }
-  ]
-  }
-}
