+++ tests/json/microjson/valid/fullexample.json
@@ -45,38 +45,5 @@
             "id": "2",
             "ref": "s3://mybucket/myfile.tif"
         }
-    ],
-    "multiscale": {
-        "axes": [
-            {
-                "name": "x",
-                "unit": "micrometer",
-                "type": "space",
-                "description": "x-axis"
-            },
-            {
-                "name": "y",
-                "unit": "micrometer",
-                "type": "space",
-                "description": "y-axis"
-            }
-        ],
-        "transformationMatrix": [
-            [
-                1.0,
-                0.0,
-                0.0
-            ],
-            [
-                0.0,
-                1.0,
-                0.0
-            ],
-            [
-                0.0,
-                0.0,
-                0.0
-            ]
-        ]
-    }
+    ]
 }
