+++ tests/json/microjson/valid/metadata-multilevel.json
@@ -95,38 +95,5 @@
     "string": {
       "plate": "plate1"
     }
-  },
-  "multiscale": {
-    "axes": [
-      {
-        "name": "x",
-        "unit": "micrometer",
-        "type": "space",
-        "description": "x-axis"
-      },
-      {
-        "name": "y",
-        "unit": "micrometer",
-        "type": "space",
-        "description": "y-axis"
-      }
-    ],
-    "transformationMatrix": [
-      [
-        1.0,
-        0.0,
-        0.0
-      ],
-      [
-        0.0,
-        1.0,
-        0.0
-      ],
-      [
-        0.0,
-        0.0,
-        0.0
-      ]
-    ]
   }
 }
