--- tests/json/microjson/valid/multilevel-coordinatesystems.json
@@ -1,82 +1,0 @@
-{
-    "type": "FeatureCollection",
-    "features": [
-        {
-            "type": "Feature",
-            "geometry": {
-                "type": "Point",
-                "coordinates": [
-                    200.0,
-                    150.0
-                ]
-            },
-            "properties": {
-                "name": "Reference Point",
-                "description": "Specific point of interest",
-                "color": "red",
-                "property1": "value1"
-            },
-            "id": "1"
-        },
-        {
-            "type": "Feature",
-            "geometry": {
-                "type": "LineString",
-                "coordinates": [
-                    [
-                        100.0,
-                        100.0
-                    ],
-                    [
-                        200.0,
-                        200.0
-                    ],
-                    [
-                        300.0,
-                        100.0
-                    ]
-                ]
-            },
-            "properties": {
-                "name": "Cell Path",
-                "description": "Path traced within a cell",
-                "color": "blue"
-            },
-            "id": "2",
-            "ref": "s3://mybucket/myfile.tif"
-        }
-    ],
-    "multiscale": {
-        "axes": [
-            {
-                "name": "x",
-                "unit": "micrometer",
-                "type": "space",
-                "description": "x-axis"
-            },
-            {
-                "name": "y",
-                "unit": "micrometer",
-                "type": "space",
-                "description": "y-axis"
-            }
-        ],
-        "tranformationMatrix": [
-            [
-                1.0,
-                0.0,
-                0.0
-            ],
-            [
-                0.0,
-                1.0,
-                0.0
-            ],
-            [
-                0.0,
-                0.0,
-                0.0
-            ]
-        ]
-    }
-}
