+++ tests/json/tilejson/invalid/empty_multiscale.json
@@ -36,36 +36,9 @@
     ],
     "fillzoom": 3,
     "multiscale": {
-        "axes": [
-            {
-                "name": "x",
-                "unit": "micrometer",
-                "type": "space",
-                "description": "x-axis"
-            },
-            {
-                "name": "y",
-                "unit": "micrometer",
-                "type": "space",
-                "description": "y-axis"
-            }
-        ],
-        "transformationMatrix": [
-            [
-                1.0,
-                0.0,
-                0.0
-            ],
-            [
-                0.0,
-                1.0,
-                0.0
-            ],
-            [
-                0.0,
-                0.0,
-                0.0
-            ]
+        "units": [
+          "micrometer",
+          "micrometer"
         ]
-    }
+      }
 }
